sozi_adm/pyproject.toml
[tool.poetry]
name = "sozi_adm"
version = "0.1.0"
description = "Sozidateli bot - Admin Panel"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
